+++ src/linemod_detect.cpp
@@ -144,8 +144,8 @@ struct Detector: public object_recognition_core::db::bases::ModelReaderBase {
       renderer_radius_min_  = document.get_field<double>("renderer_radius_min");
       renderer_radius_max_ = document.get_field<double>("renderer_radius_max");
       renderer_radius_step_ = document.get_field<double>("renderer_radius_step");
-      renderer_width_ = document.get_field<size_t>("renderer_width");
-      renderer_height_ = document.get_field<size_t>("renderer_height");
+      renderer_width_ = document.get_field<int>("renderer_width");
+      renderer_height_ = document.get_field<int>("renderer_height");
       renderer_focal_length_x_ = document.get_field<double>("renderer_focal_length_x");
       renderer_focal_length_y_ = document.get_field<double>("renderer_focal_length_y");
       renderer_near_ = document.get_field<double>("renderer_near");
@@ -520,9 +520,9 @@ struct Detector: public object_recognition_core::db::bases::ModelReaderBase {
     /** Renderer parameter: the step scale sampling*/
     double renderer_radius_step_;
     /** Renderer parameter: image width */
-    size_t renderer_width_;
+    int renderer_width_;
     /** Renderer parameter: image height */
-    size_t renderer_height_;
+    int renderer_height_;
     /** Renderer parameter: near distance */
     double renderer_near_;
     /** Renderer parameter: far distance */
